requirements.txt
astroid==2.3.3
autopep8==1.5
azure-cognitiveservices-vision-customvision==1.0.0
azure-common==1.1.24
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.9
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msrest==0.6.11
oauthlib==3.1.0
pkg-resources==0.0.0
pycodestyle==2.5.0
pylint==2.4.4
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==0.16.1
wrapt==1.11.2

